requirements.txt
openai
langchain
langchain-community
faiss-cpu
flask
python-dotenv
tiktoken